# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/passengers": {
      "post": {
        "operationId": "PassengersController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePassengerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Passenger created successfully"
          },
          "400": {
            "description": "Validation error (must have exactly one owner)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a passenger",
        "tags": [
          "passengers"
        ]
      },
      "get": {
        "operationId": "PassengersController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all passengers"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all passengers (Admin/Superuser only)",
        "tags": [
          "passengers"
        ]
      }
    },
    "/passengers/{id}": {
      "get": {
        "operationId": "PassengersController_findOne",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the passenger"
          },
          "403": {
            "description": "Forbidden (not owner)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a passenger by ID (ownership checked via Guard)",
        "tags": [
          "passengers"
        ]
      },
      "patch": {
        "operationId": "PassengersController_update",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePassengerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Passenger updated"
          },
          "403": {
            "description": "Forbidden (not owner)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a passenger (ownership checked via Guard)",
        "tags": [
          "passengers"
        ]
      },
      "delete": {
        "operationId": "PassengersController_remove",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Passenger deleted"
          },
          "403": {
            "description": "Forbidden (not owner)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a passenger (ownership checked via Guard)",
        "tags": [
          "passengers"
        ]
      }
    },
    "/passengers/traveler/{travelerId}": {
      "get": {
        "operationId": "PassengersController_findByTraveler",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "travelerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return traveler passengers"
          },
          "403": {
            "description": "Forbidden (not owner)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get passengers by traveler ID (ownership checked via Guard)",
        "tags": [
          "passengers"
        ]
      }
    },
    "/passengers/agency/{agencyId}": {
      "get": {
        "operationId": "PassengersController_findByAgency",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "description": "User role for RBAC. Allowed values: traveler, guide, support, admin, superuser, agency",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "traveler",
                "guide",
                "support",
                "admin",
                "superuser",
                "agency"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for ownership verification",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agencyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return agency passengers"
          },
          "403": {
            "description": "Forbidden (not owner)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get passengers by agency ID (ownership checked via Guard)",
        "tags": [
          "passengers"
        ]
      }
    }
  },
  "info": {
    "title": "GlobeSync API",
    "description": "GlobeSync Travel Platform Backend API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          },
          "message": {
            "type": "object",
            "example": "Resource with the given ID was not found"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-05T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/travelers/abc-123"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ]
      },
...
      "CreatePassengerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "age": {
            "type": "number",
            "example": 30
          },
          "gender": {
            "type": "string",
            "example": "Female"
          },
          "travelerId": {
            "type": "number",
            "example": 1
          },
          "agencyId": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "name",
          "age",
          "gender"
        ]
      },
      "UpdatePassengerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "age": {
            "type": "number",
            "example": 30
          },
          "gender": {
            "type": "string",
            "example": "Female"
          },
          "travelerId": {
            "type": "number",
            "example": 1
          },
          "agencyId": {
            "type": "number",
            "example": 1
          }
        }
      }
    }
  }
}